+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   ([#479](https://github.com/shioyama/mobility/pull/479))
 - Add `Mobility.validate_locale!`
   ([#479](https://github.com/shioyama/mobility/pull/479))
+- Include query_scope in `find_by_` methods
+  ([#480](https://github.com/shioyama/mobility/pull/480))
 
 ## 1.0
 
@@ -20,7 +22,7 @@ for more details on how to upgrade.
 
 - Make `Mobility::Plugins::ActiveRecord::Query::VirtualRow` and
   `Mobility::Plugins::ActiveRecord::Query::QueryExtension`
-  ([#471](https://github.com/shioyama/mobility/pull/471)) public
+  public ([#471](https://github.com/shioyama/mobility/pull/471))
 - Fix typo in initializer template
   ([#474](https://github.com/shioyama/mobility/pull/474))